Line 20 in components/Head.tsx was added to accept the image string URL. As a result the constant, PageMetaProps, at line 31 of pages/quest/[slug].tsx was modified to also handle that string and pass all three desired metadata properties to the Head component. In order to store the filtered metadata from constants/quests.ts in constants/index.ts I added a new type at the bottom of entities/quest.ts. Finally in constants/index.ts from line 28 to the end is the function that iterates over each quest, filtering out the three desired metadata properties, creating a new array of objects that is assigned to the constant QUEST_METADATA.
NOTE: Because the variable "currentQuest" in /pages/quest/[slug].tsx already contains all of the data that is stored in QUEST_METADATA I simply passed the desired metadata properties from currentQuest to pageMeta instead.
This saves another round trip of having to find the correct object as was already done when currentQuest was defined.