Deleted the console.log() left at the bottom of constants/index.ts Had to add a "image: '...'" line to each of the PageMetaProps constants for each page to fit the new type format from the last commit.
Reading more about the way Next does SSR it appears that the project is still using purely static site generation.
In order to implement SSR the following must be used "xport async function getServerSideProps()" but does not appear to be present anywhere in the project.
Static generation is done using "export async function getStaticProps()" which is present.